Minor fixes : Windows compiling issues, warnings because of casts and whitespaces
git-svn-id: http://bob.alpha.agorabox.org/repos/glumol/trunk/poujol@285 29d5ca58-9003-0410-8356-a7fefff29482

+++ build_animation.cpp
@@ -36,7 +36,7 @@ BOOST_PYTHON_MEMBER_FUNCTION_OVERLOADS(
 BOOST_PYTHON_MEMBER_FUNCTION_OVERLOADS(
 	add_frame_overloads, add_frame, 1, 3)
 
-float get_integer(object obj)
+int get_integer(object obj)
 {
     if(PyObject_IsInstance(obj.ptr(), (PyObject*) &PyFloat_Type))
         return (int) extract<float>(obj);
